refactor(dsync): overhaul dsync API around push/pull & transfer
dsync's api is, well bad. Renaming things to make it clearer. This commit breaks the api,
I'll follow up with docs cleanup & some additions to round out the way dsync works

BREAKING CHANGE:
api is completely overhauled dependants will need to refactor
