+++ dsync/push.go
@@ -10,15 +10,15 @@ import (
 	ipld "github.com/ipfs/go-ipld-format"
 )
 
-// NewSend initiates a send for a DAG at an id from a local to a remote.
-// Send is initiated by the local node
-func NewSend(ctx context.Context, lng ipld.NodeGetter, mfst *dag.Manifest, remote Remote) (*Send, error) {
-	parallelism := defaultSendParallelism
+// NewPush initiates a send for a DAG at an id from a local to a remote.
+// Push is initiated by the local node
+func NewPush(ctx context.Context, lng ipld.NodeGetter, mfst *dag.Manifest, remote Remote) (*Push, error) {
+	parallelism := defaultPushParallelism
 	if len(mfst.Nodes) < parallelism {
 		parallelism = len(mfst.Nodes)
 	}
 
-	ps := &Send{
+	ps := &Push{
 		ctx:         ctx,
 		mfst:        mfst,
 		lng:         lng,
@@ -32,8 +32,8 @@ func NewSend(ctx context.Context, lng ipld.NodeGetter, mfst *dag.Manifest, remot
 	return ps, nil
 }
 
-// Send coordinates requesting & sending a manifest to a remote, tracking progress and state
-type Send struct {
+// Push coordinates requesting & sending a manifest to a remote, tracking progress and state
+type Push struct {
 	sid         string          // session ID for this push, generated by remote
 	ctx         context.Context // session context
 	mfst        *dag.Manifest   // manifest we're sending
@@ -49,7 +49,7 @@ type Send struct {
 }
 
 // Do executes the send, blocking until complete
-func (snd *Send) Do() (err error) {
+func (snd *Push) Do() (err error) {
 	// First Do sends a manifest to the remote node
 	// The remote returns the id of the process, and a diff of the manifest
 	// which describes which blocks we need to send
@@ -71,22 +71,22 @@ func (snd *Send) Do() (err error) {
 	// posible TODO (ramfox): it would be great if the fetch and send Do functions
 	// followed the same pattern. Specifically the go function that is used to listen for
 	// responses
-	snd.sid, snd.diff, err = snd.remote.ReqSend(snd.mfst)
+	snd.sid, snd.diff, err = snd.remote.PushStart(snd.mfst)
 	if err != nil {
 		return err
 	}
-	return snd.doActualSend()
+	return snd.doActualPush()
 }
 
-// PerformSend executes the send, given that we already have a session and diff
-func (snd *Send) PerformSend(sid string, mfst, diff *dag.Manifest) error {
+// Push executes the send, given that we already have a session and diff
+func (snd *Push) Push(sid string, mfst, diff *dag.Manifest) error {
 	snd.sid = sid
 	snd.mfst = mfst
 	snd.diff = diff
-	return snd.doActualSend()
+	return snd.doActualPush()
 }
 
-func (snd *Send) doActualSend() (err error) {
+func (snd *Push) doActualPush() (err error) {
 	snd.prog = dag.NewCompletion(snd.mfst, snd.diff)
 	go snd.completionChanged()
 	// defer close(snd.progCh)
@@ -172,11 +172,11 @@ func (snd *Send) doActualSend() (err error) {
 }
 
 // Completion returns a read-only channel of updates to completion
-func (snd *Send) Completion() <-chan dag.Completion {
+func (snd *Push) Completion() <-chan dag.Completion {
 	return snd.progCh
 }
 
-func (snd *Send) completionChanged() {
+func (snd *Push) completionChanged() {
 	snd.progCh <- snd.prog
 }
 
@@ -219,7 +219,7 @@ func (s sender) start() {
 					}
 					return
 				}
-				s.responses <- s.remote.PutBlock(s.sid, hash, node.RawData())
+				s.responses <- s.remote.PushBlock(s.sid, hash, node.RawData())
 			}()
 		case <-s.stopCh:
 			return
